progressed with the demo notebooks; some minor fixes and changes
